requirements.txt
Pillow
numpy
matplotlib

